# docs/swagger.yaml
basePath: /api/v1
definitions:
  user.AttachmentDTO:
    properties:
      file_name:
        example: document.jpg
        type: string
      file_path:
        example: /assets/images/20240115_a1b2c3d4.jpg
        type: string
    required:
    - file_name
    - file_path
    type: object
  user.DepartmentInfo:
    properties:
      address:
        example: 123 Đường ABC, Quận 1, TP.HCM
        type: string
      code:
        example: SYT
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Sở Y tế
        type: string
    type: object
  user.LoginRequest:
    properties:
      citizen_id:
        example: "001234567890"
        type: string
      password:
        example: Aa@123456
        type: string
    required:
    - citizen_id
    - password
    type: object
  user.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/user.UserInfo'
    type: object
  user.MyAppItemResponse:
    properties:
      code:
        example: HS-20240115-a1b2c3d4
        type: string
      created_at:
        example: "2024-01-15T10:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      service_name:
        example: Đăng ký kinh doanh
        type: string
      status:
        example: Received
        type: string
    type: object
  user.MyAppListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/user.MyAppItemResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_items:
        example: 20
        type: integer
      total_pages:
        example: 2
        type: integer
    type: object
  user.ProfileResponse:
    properties:
      address:
        example: 123 Đường ABC, Quận 1, TP.HCM
        type: string
      citizen_id:
        example: "001234567890"
        type: string
      created_at:
        example: "2024-01-01T10:00:00Z"
        type: string
      date_of_birth:
        example: "1990-01-15"
        type: string
      email:
        example: nguyenvana@example.com
        type: string
      gender:
        example: male
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_email_notify:
        example: true
        type: boolean
      name:
        example: Nguyễn Văn A
        type: string
      phone:
        example: "0901234567"
        type: string
      role:
        example: citizen
        type: string
    type: object
  user.RegisterRequest:
    properties:
      address:
        example: 123 Đường ABC, Quận 1, TP.HCM
        type: string
      citizen_id:
        example: "001234567890"
        type: string
      date_of_birth:
        example: "1990-01-15"
        type: string
      email:
        example: nguyenvana@example.com
        type: string
      gender:
        enum:
        - male
        - female
        - other
        example: male
        type: string
      name:
        example: Nguyễn Văn A
        type: string
      password:
        example: Aa@123456
        type: string
      phone:
        example: "0901234567"
        type: string
    required:
    - citizen_id
    - email
    - gender
    - name
    - password
    - phone
    type: object
  user.ServiceDetailResponse:
    properties:
      code:
        example: SV001
        type: string
      created_at:
        example: "2024-01-01T10:00:00Z"
        type: string
      department:
        $ref: '#/definitions/user.DepartmentInfo'
      description:
        example: Dịch vụ đăng ký kinh doanh cho cá nhân và doanh nghiệp
        type: string
      fee:
        example: 500000
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Đăng ký kinh doanh
        type: string
      processing_days:
        example: 5
        type: integer
      sector:
        example: Health
        type: string
    type: object
  user.ServiceListItem:
    properties:
      code:
        example: SV001
        type: string
      department_name:
        example: Sở Y tế
        type: string
      fee:
        example: 500000
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Đăng ký kinh doanh
        type: string
      processing_days:
        example: 5
        type: integer
      sector:
        example: Health
        type: string
    type: object
  user.ServiceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/user.ServiceListItem'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_items:
        example: 50
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  user.SubmitAppRequest:
    properties:
      attachments:
        items:
          $ref: '#/definitions/user.AttachmentDTO'
        type: array
      service_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - attachments
    - service_id
    type: object
  user.SubmitAppResponse:
    properties:
      code:
        example: HS-20240115-a1b2c3d4
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  user.UpdateProfileRequest:
    properties:
      address:
        example: 123 Đường ABC, Quận 1, TP.HCM
        type: string
      date_of_birth:
        example: "1990-01-15"
        type: string
      gender:
        enum:
        - male
        - female
        - other
        example: male
        type: string
      is_email_notify:
        example: true
        type: boolean
      name:
        example: Nguyễn Văn A
        type: string
      phone:
        example: "0901234567"
        type: string
    required:
    - gender
    - name
    - phone
    type: object
  user.UploadFileResponse:
    properties:
      file_name:
        example: 20240115143022_a1b2c3d4.jpg
        type: string
      file_size:
        example: 1024000
        type: integer
      file_url:
        example: /assets/images/20240115143022_a1b2c3d4.jpg
        type: string
      original_name:
        example: document.jpg
        type: string
    type: object
  user.UserInfo:
    properties:
      citizen_id:
        example: "001234567890"
        type: string
      email:
        example: nguyenvana@example.com
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Nguyễn Văn A
        type: string
      phone:
        example: "0901234567"
        type: string
      role:
        example: citizen
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      status:
        example: 200
        type: integer
    type: object
  utils.ValidationErrors:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        example: Validation failed
        type: string
      status:
        example: 400
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API cho hệ thống quản lý dịch vụ công
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Public Service Management API
  version: "1.0"
paths:
  /applications:
    post:
      consumes:
      - application/json
      description: Công dân nộp hồ sơ dịch vụ công
      parameters:
      - description: Thông tin hồ sơ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.SubmitAppRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.SubmitAppResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Nộp hồ sơ
      tags:
      - Applications
  /applications/me:
    get:
      consumes:
      - application/json
      description: Lấy danh sách hồ sơ đã nộp của công dân
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - Received
        - Processing
        - Supplement_Required
        - Approved
        - Rejected
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.MyAppListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Danh sách hồ sơ của tôi
      tags:
      - Applications
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập bằng CCCD và mật khẩu
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Đăng nhập
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Đăng xuất khỏi hệ thống (client xóa token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Đăng xuất
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Đăng ký tài khoản công dân mới
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.UserInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrors'
      summary: Đăng ký tài khoản
      tags:
      - Auth
  /profile:
    get:
      consumes:
      - application/json
      description: Lấy thông tin hồ sơ công dân
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.ProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin hồ sơ
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin hồ sơ công dân
      parameters:
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.ProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật hồ sơ
      tags:
      - Profile
  /services:
    get:
      consumes:
      - application/json
      description: Lấy danh sách dịch vụ với pagination và filter
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or code
        in: query
        name: keyword
        type: string
      - description: Filter by sector
        in: query
        name: sector
        type: string
      - description: Filter by department ID
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.ServiceListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Danh sách dịch vụ
      tags:
      - Services
  /services/{id}:
    get:
      consumes:
      - application/json
      description: Lấy thông tin chi tiết dịch vụ
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.ServiceDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Chi tiết dịch vụ
      tags:
      - Services
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file (images, documents) lên server
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.UploadFileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "429":
          description: Too Many Requests - respect Retry-After header
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying the upload
              type: string
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload file
      tags:
      - Upload
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"